Schmitt decomposition now method of StateVector
