+++ quasim/src/ext.rs
@@ -317,85 +317,6 @@ pub fn expand_matrix(
     sum
 }
 
-/// # coefficent_matrix
-/// Finds the appropriate coefficent matrix
-/// used in schmitt decomposition.
-fn coefficent_matrix(
-    state: &impl Index<usize, Output = Complex<f64>>,
-    targets: &[usize],
-    n_qubits: usize,
-) -> DMatrix<Complex<f64>> {
-    let squash_by_mask = |bitstring: usize, mask: usize| {
-        let mut res = 0;
-        let mut i = 0;
-        let mut j = 0;
-
-        while (mask >> i) != 0 {
-            if (mask >> i) & 1 != 0 {
-                res |= ((bitstring >> i) & 1) << j;
-                j += 1;
-            }
-            i += 1;
-        }
-
-        res
-    };
-
-    let n_elements = 1 << n_qubits;
-    let n_targets = targets.len();
-
-    let mut coeffs = DMatrix::<Complex<f64>>::zeros(n_elements >> n_targets, 1 << n_targets);
-
-    let target_mask = QBits::from_indices(targets).get_bitstring();
-    let not_target_mask = (n_elements - 1) ^ target_mask;
-
-    for i in 0..n_elements {
-        let col = squash_by_mask(i, target_mask);
-        let row = squash_by_mask(i, not_target_mask);
-        coeffs[(row, col)] = state[i];
-    }
-
-    coeffs
-}
-
-/// # schmitt_trace
-/// Assuming state can be factored,
-/// returns the state with targets traced out.
-pub fn schmitt_trace(
-    state: &impl Index<usize, Output = Complex<f64>>,
-    targets: &[usize],
-    n_qubits: usize,
-) -> DVector<Complex<f64>> {
-    // Find the coefficent matrix for the state vector.
-    let coeffs = coefficent_matrix(state, targets, n_qubits);
-
-    // Use Singular Value Decomposition to find the traced out vector.
-    let svd = coeffs.svd(true, false);
-
-    let res = DVector::<Complex<f64>>::from(svd.u.unwrap().column(0));
-    // SVD might mess with global phase.
-    res.scale(res[0].re.signum())
-}
-
-/// # schmitt_reduce
-/// Assuming state can be factored,
-/// returns the reduced state of targets.
-pub fn schmitt_reduce(
-    state: &impl Index<usize, Output = Complex<f64>>,
-    targets: &[usize],
-    n_qubits: usize,
-) -> DVector<Complex<f64>> {
-    // Find the coefficent matrix for the state vector.
-    let coeffs = coefficent_matrix(state, targets, n_qubits);
-
-    // Use Singular Value Decomposition to find the traced out vector.
-    let svd = coeffs.svd(false, true);
-
-    let res = DVector::<Complex<f64>>::from(svd.v_t.unwrap().transpose().column(0));
-    // SVD might mess with global phase.
-    res.scale(res[0].re.signum())
-}
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct SortedVec<T, K = T>(Vec<T>, PhantomData<K>);
 impl<T: OrdByKey<K>, K: Ord> SortedVec<T, K> {
