+++ quasim/src/prod_simulator.rs
@@ -1,7 +1,7 @@
 use crate::{
     circuit::{Circuit, HybridCircuit, PureCircuit, pc::CircuitPc},
     expr_dsl::{BitExpr, BoolExpr},
-    ext::{collapse, measure_state_vector, schmitt_trace},
+    ext::{collapse, measure_state_vector},
     gate::{Gate, GateType},
     instruction::Instruction,
     product_state::{ProductState, SubSystem},
@@ -174,11 +174,10 @@ impl ProdSimulator {
 
         // "Split" state.
         sys.qubits_mut().remove(local_target);
-        *sys.state_vector_mut() = StateVector::from(schmitt_trace(
-            sys.state_vector(),
+        *sys.state_vector_mut() = sys.state_vector().schmitt_trace(
             &[local_target],
             local_n_qubits,
-        ));
+        );
 
         self.product_state[target_system] = sys;
